Final Audio and Architectural Enhancements
Address all remaining outstanding issues from the analysis:
- Fix Arpeggiator overflow potential by changing stepIndex_ to int16_t.
- Implement overshoot technique in Envelope class for accurate timing.
- Implement supersaw detune scaling and coarse tuning support.
- Extend Preset system to version 2, covering all sound parameters.
- Implement MIDI output for sequencers and "Legato Only" glide mode.
- Add Reverb Freeze mode and fix MIDI running status for system messages.
- Further optimize signal chain with panning caching and better effect mixing.
- Ensure full stereo integrity and thread safety across the engine.

+++ Arpeggiator.h
@@ -73,7 +73,7 @@ private:
     uint8_t sortedVelocities_[ARP_MAX_NOTES];
     
     // Playback state
-    int8_t stepIndex_;
+    int16_t stepIndex_;
     int8_t direction_;  // 1 or -1
     uint8_t currentOctave_;
     uint8_t currentNote_;
